requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
pycodestyle==2.7.0
PyJWT==1.4.2
pymongo==3.12.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
